restructured folders in preparation for allocator abstraction.
