+++ src/allocator/result.rs
